% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_eca.R
\name{climdex.cwd.eobs}
\alias{climdex.cwd.eobs}
\title{CONSECUTIVE WET DAYS
We would This is a trial to have cwd for months also (not sure about this)
 spells.can.span.years=FALSE}
\usage{
climdex.cwd.eobs(ci, freq = c("monthly", "annual"),
  spells.can.span.years = FALSE)
}
\description{
CONSECUTIVE WET DAYS
We would This is a trial to have cwd for months also (not sure about this)
 spells.can.span.years=FALSE
}

