% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdex.r
\name{climdex.HI}
\alias{climdex.HI}
\title{Huglin Index}
\usage{
climdex.HI(ci, freq = c("annual"), cur_sub)
}
\arguments{
\item{ci}{Object of type climdexInput (representing the daily mean and daily max temperature)}

\item{freq}{Time frequency to aggregate to. Allowed are only "annual"}

\item{cur_sub}{TO DO!}

\item{unit.}{Allowed are only deg C.}
}
\value{
A vector containing the HI
}
\description{
This function computes the climdex index HI:
}
\author{
Christiana Photiadou (KNMI)
}
\references{
\url{http://www.ecad.eu/indicesextremes/indicesdictionary.php}
}

