% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdex.r
\name{climdex.sd}
\alias{climdex.sd}
\title{Mean snow depth}
\usage{
climdex.sd(ci, freq = c("annual", "monthly", "halfyear", "seasonal"))
}
\arguments{
\item{ci}{Object of type climdexInput (representing the daily snow depth timeseries in [cm]).}

\item{freq}{Time frequency to aggregate to. Allowed are: "annual","halfyear" or "seasonal". Default: "annual".}
}
\value{
A vector containing the timeseries of the mean snow depth.
}
\description{
This function computes the climdex index SD: The mean snow depth measured within a period.
}
\author{
Rebekka Posselt (MeteoSwiss)
}
\references{
\url{http://www.ecad.eu/indicesextremes/indicesdictionary.php}
}

