% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdex.r
\name{climdex.sun_sunny}
\alias{climdex.sun_sunny}
\title{Mostly sunny days}
\usage{
climdex.sun_sunny(ci, freq = c("annual", "halfyear", "seasonal", "monthly"))
}
\arguments{
\item{ci}{Object of type climdexInput (representing the daily relative sunshine duration in [\%])}

\item{freq}{Time frequency to aggregate to. Allowed are: "annual","halfyear", "seasonal" or "monthly". Default: "annual".}
}
\value{
A vector containing the number of mostly sunny days.
}
\description{
This function computes the climdex index SUN_SUNNY.
}
\details{
This function takes a climdexInput object as input and computes the
SUN_SUNNY index: The number of mostly sunny days (relative sunshine duration > 80 \%) within a period.
}
\references{
\url{http://etccdi.pacificclimate.org/list_27_indices.shtml}
}
\seealso{
\code{\link{climdexInput.raw}}, \code{\link{climdexInput.csv}}.
}
\keyword{climate}
\keyword{ts}

