% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdex.r, R/indices_eca.R
\name{curry_in_subset_for_huglin}
\alias{curry_in_subset_for_huglin}
\title{Introduces by C.Photiadou (KNMI)}
\usage{
curry_in_subset_for_huglin(cdx.funcs, subset)

curry_in_subset_for_huglin(cdx.funcs, subset)
}
\description{
Introduces by C.Photiadou (KNMI)

HUGLIN INDEX
This function is not ready yet. It is uses coeeficient based on the latitude
For this index I had to curry the cdx.funcs to be able to include the subset. Later I realised 
I need also to include the latitude. This would be used in compute.indices.for.stripe together with get.lat 
to retrieve subset & latitude 
I didn't proceed with finisheing the eca.HI function. I thought to ask you first if its possible 
to adapt compute.indices.for.stripe so it can include the currying and the latitude. Or if you had a better idea on this 
please let me know. 
Function to Curry a cxd.funcs for subset
used only for Huglin Index
}

