% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/climdex.pcic-package.R
\docType{data}
\name{ec.1018935}
\alias{ec.1018935}
\alias{ec.1018935.prec}
\alias{ec.1018935.tmax}
\alias{ec.1018935.tmin}
\title{EC example data}
\format{A data frame consisting of year, day of year, precipitation or daily
maximum temperature or daily minimum temperature on that day, and
observation flags.}
\description{
This is the Environment Canada CDCD (Canadian Daily Climate Data)
precipitation, maximum temperature, and minimum temperature data for station
1018935 - William Head, BC, Canada.
}
\details{
This is the Environment Canada CDCD (Canadian Daily Climate Data)
precipitation, daily maximum temperature, and daily minimum temperature data
for station 1018935 - William Head, BC, Canada. This is provided as example
data for running the Climdex package.
}
\seealso{
\code{\link{climdexInput.raw}},
http://www.climate.weatheroffice.gc.ca/prods_servs/index_e.html#cdcd .
}

