% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_eca.R
\name{eca.cfd}
\alias{eca.cfd}
\title{CONSECUTIVE FROST DAYS
maximum number of concsecutive frost days (tn < 0C)
spell.can.span.years for us is FALSE}
\usage{
eca.cfd(ci, freq = c("monthly", "annual"), spells.can.span.years = FALSE)
}
\description{
CONSECUTIVE FROST DAYS
maximum number of concsecutive frost days (tn < 0C)
spell.can.span.years for us is FALSE
}

