% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_eca.R
\name{eca.csu}
\alias{eca.csu}
\title{CONSECUTIVE SUMMER DAYS
maximum number of concsecutive summer days (tx > 25C)
spell.can.span.years for us is FALSE}
\usage{
eca.csu(ci, freq = c("monthly", "annual"), spells.can.span.years = FALSE)
}
\description{
CONSECUTIVE SUMMER DAYS
maximum number of concsecutive summer days (tx > 25C)
spell.can.span.years for us is FALSE
}

