% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_eca.R
\name{eca.HI}
\alias{eca.HI}
\title{Huglin index 
It is annual only and valid for months (April-Sep)}
\usage{
eca.HI(ci, freq = c("annual"), subset = -1)
}
\description{
Huglin index 
It is annual only and valid for months (April-Sep)
}

