% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_eca.R
\name{eca.SPI3}
\alias{eca.SPI3}
\title{In ECA&D if 85% of data are present then we consider them as valid. 
This was not used in SPI or elsewhere (yet)
empirical Standard Precipitation Index 3 (k in getPrecOnTimescale is 3)}
\usage{
eca.SPI3(ci, freq = c("monthly"))
}
\description{
In ECA&D if 85% of data are present then we consider them as valid. 
This was not used in SPI or elsewhere (yet)
empirical Standard Precipitation Index 3 (k in getPrecOnTimescale is 3)
}

