% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_eca.R
\name{eca.SPI6}
\alias{eca.SPI6}
\title{empirical Standard Precipitation Index 6 (k in getPrecOnTimescale is 6)}
\usage{
eca.SPI6(ci, freq = c("monthly"))
}
\description{
empirical Standard Precipitation Index 6 (k in getPrecOnTimescale is 6)
}

