% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climdex_function_eobs.R
\name{get.climdex.functions.eobs}
\alias{get.climdex.functions.eobs}
\title{Returns a list of Climdex functions, with parameters curried in.
Create func.names.eca, options.eca and func.eca}
\usage{
get.climdex.functions.eobs(vars.list, fclimdex.compatible = TRUE)
}
\description{
Returns a list of Climdex functions, with parameters curried in.
Create func.names.eca, options.eca and func.eca
}

