% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climdex_function_eobs.R
\name{get.climdex.variable.list.eobs}
\alias{get.climdex.variable.list.eobs}
\title{Returns a list of Climdex variables given constraints
Added: monthly only for SPI indices
Changed annual only is changed 
Added huglin index in tavg although it takes two input variables (tmax and tavg)
Changed fre.list to accept also only months and added helper_fun for same reason}
\usage{
get.climdex.variable.list.eobs(source.data.present, time.resolution = c("all",
  "annual", "monthly"), climdex.vars.subset = NULL)
}
\description{
Returns a list of Climdex variables given constraints
Added: monthly only for SPI indices
Changed annual only is changed 
Added huglin index in tavg although it takes two input variables (tmax and tavg)
Changed fre.list to accept also only months and added helper_fun for same reason
}

