% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_climdex_function_eobs.R
\name{get.climdex.variable.metadata.eobs}
\alias{get.climdex.variable.metadata.eobs}
\title{Returns metadata for specified Climdex variables
Add details for additional indices}
\usage{
get.climdex.variable.metadata.eobs(vars.list, template.filename)
}
\description{
Returns metadata for specified Climdex variables
Add details for additional indices
}

