% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/climdex.r
\name{get.last.monthday.of.year}
\alias{get.last.monthday.of.year}
\title{Get the last month and day of the year}
\usage{
get.last.monthday.of.year(d, sep = "-")
}
\arguments{
\item{d}{An exemplar date.}

\item{sep}{Separator to use.}
}
\value{
A string (like "12-30", or "12-31")
}
\description{
Get the last month and day of the year as a character sting, separated by
the specified separator.
}
\details{
This is a utility function necessitated by 360-day calendars. Works on PCICt objects.
}
\examples{
library(PCICt)
last.mday <- get.last.monthday.of.year(as.PCICt("2011-01-01", cal="360"))
}

