% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_eca.R
\name{get.lat}
\alias{get.lat}
\title{Function to retrieve the latitude 
 we have different variable names for temp & precip; that's why I've changed the variable.name.map}
\usage{
get.lat(open_file_list, v.f.idx, variable.name.map = c(tmax = "tx", tmin =
  "tn", prec = "rr", tavg = "tg"))
}
\description{
Function to retrieve the latitude 
 we have different variable names for temp & precip; that's why I've changed the variable.name.map
}

