% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices_eca.R
\name{monthly_sums_spi}
\alias{monthly_sums_spi}
\title{STANDARDIZE PRECIPITATION INDEX
based on the formula in the spi package 
SPI is only has monthly outputs (see changes in get.climdex.variable.list.eobs)
make monthly sums}
\usage{
monthly_sums_spi(temp, date.factor)
}
\description{
STANDARDIZE PRECIPITATION INDEX
based on the formula in the spi package 
SPI is only has monthly outputs (see changes in get.climdex.variable.list.eobs)
make monthly sums
}

