% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdex.r
\name{nday.consec.temp.mean}
\alias{nday.consec.temp.mean}
\title{Number of days (less than, greater than, etc) a threshold}
\usage{
nday.consec.temp.mean(daily.temp, date.factor, ndays, freq.fun = "max",
  center.mean.on.last.day = FALSE)
}
\arguments{
\item{daily.temp}{Daily timeseries of temperature (tmax, tavg or tmin).}

\item{date.factor}{Factor to aggregate by.}

\item{ndays}{Number of days in the running window.}

\item{freq.fun}{Function to apply on the choosen period (e.g., max, min, ...)}

\item{center.mean.on.last.day}{Whether to center the n-day running mean on
the last day of the series, instead of the middle day.}
}
\value{
A vector consisting of the mean n-day temp per
time interval.
}
\description{
Produces means of values that exceed (or are below) the specified threshold.
}
\details{
This function takes a data series, the number of days in the running window,
a date factor to aggregate by, and an optional modifier parameter
(center.mean.on.last.day). It computes the n-day running mean of
temperature and returns the mean n-day temperature per unit
time, as defined by \code{date.factor}.
}
\examples{
library(PCICt)

## Parse the dates into PCICt.
tmax.dates <- as.PCICt(do.call(paste, ec.1018935.tmax[,c("year","jday")]),
                       format="\%Y \%j", cal="gregorian")
tmin.dates <- as.PCICt(do.call(paste, ec.1018935.tmin[,c("year","jday")]),
                       format="\%Y \%j", cal="gregorian")
prec.dates <- as.PCICt(do.call(paste, ec.1018935.prec[,c("year","jday")]),
                       format="\%Y \%j", cal="gregorian")

## Load the data in.
ci <- climdexInput.raw(tmax=ec.1018935.tmax$MAX_TEMP,
                       tmin=ec.1018935.tmin$MIN_TEMP,
                       prec=ec.1018935.prec$ONE_DAY_PRECIPITATION,
                       tmax.dates=tmax.dates,
                       tmin.dates=tmin.dates,
                       prec.dates=prec.dates,
                       base.range=c(1971, 2000))

## Compute txnday on a monthly basis.
tx5daymax <- nday.consec.temp.mean(ci@data$tmax, ci@date.factors$monthly, freq.fun="max", ndays=5)

}
\keyword{climate}
\keyword{ts}

