% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climdex.r
\name{running.mean}
\alias{running.mean}
\title{Running Mean of a Vector}
\usage{
running.mean(vec, bin)
}
\arguments{
\item{vec}{A vector}

\item{bin}{The number of entries to average over for each mean}
}
\value{
a vector containing the running mean of bin elements of vec
}
\description{
Calculates the running means of a vector with a shifting window
}
\details{
Returns a new vector the same length as vec, where the ith
element is the mean of the bin of elements centered at the ith element
of the original vector. Means cannot be calculated for elements less
than half the width of the bin from the beginning or end of the vector;
the result vector has NA in those positions.
}
